feat: support floating point frequency values

+++ main.go
@@ -30,7 +30,7 @@ const (
 type CompileRequest struct {
 	Sources   map[string]string `json:"sources"`
 	TopModule string            `json:"topModule"`
-	Freq      *int              `json:"freq,omitempty"`
+	Freq      *float64          `json:"freq,omitempty"`
 	Seed      *int              `json:"seed,omitempty"`
 }
 
@@ -258,7 +258,7 @@ func handleCompile(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// Run nextpnr-ice40
-	freq := 12
+	freq := 12.0
 	if req.Freq != nil {
 		freq = *req.Freq
 	}
@@ -270,7 +270,7 @@ func handleCompile(w http.ResponseWriter, r *http.Request) {
 	nextpnrArgs := []string{
 		"--pcf-allow-unconstrained",
 		"--seed", fmt.Sprintf("%d", seed),
-		"--freq", fmt.Sprintf("%d", freq),
+		"--freq", fmt.Sprintf("%.3f", freq),
 		"--package", "sg48",
 		"--up5k",
 		"--asc", "output.asc",
